TIP-856: Move Pim\ClassificationBundle classes to Akeneo\Tool
