+++ src/Akeneo/Bundle/ClassificationBundle/Resources/config/repositories.yml
@@ -1,5 +1,5 @@
 parameters:
-    akeneo_classification.repository.abstract_item_category.class: Akeneo\Bundle\ClassificationBundle\Doctrine\ORM\Repository\AbstractItemCategoryRepository
+    akeneo_classification.repository.abstract_item_category.class: Akeneo\Tool\Bundle\ClassificationBundle\Doctrine\ORM\Repository\AbstractItemCategoryRepository
 
 services:
     akeneo_classification.repository.abstract_item_category:
