+++ src/Pim/Bundle/CatalogBundle/Doctrine/ORM/Repository/ProductCategoryRepository.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository;
 
-use Akeneo\Bundle\ClassificationBundle\Doctrine\ORM\Repository\AbstractItemCategoryRepository;
+use Akeneo\Tool\Bundle\ClassificationBundle\Doctrine\ORM\Repository\AbstractItemCategoryRepository;
 use Doctrine\ORM\EntityManager;
 use Pim\Component\Catalog\Repository\ProductCategoryRepositoryInterface;
