+++ src/Pim/Bundle/CatalogBundle/Resources/config/doctrine/Category.orm.yml
@@ -2,7 +2,7 @@ Pim\Bundle\CatalogBundle\Entity\Category:
     type: entity
     table: pim_catalog_category
     changeTrackingPolicy: DEFERRED_EXPLICIT
-    repositoryClass: Akeneo\Bundle\ClassificationBundle\Doctrine\ORM\Repository\CategoryRepository
+    repositoryClass: Akeneo\Tool\Bundle\ClassificationBundle\Doctrine\ORM\Repository\CategoryRepository
     uniqueConstraints:
         pim_category_code_uc:
             columns:
